+++ Changelog.md
@@ -18,6 +18,8 @@
 - Fix: Vanilla Hoppers and Reinforced Hoppers from different owners can take out items from a Reinforced Hopper
 - Fix: A Portable Radar does not send a message to its owner if the owner is not part of a team
 - Fix: Adding an empty Disguise Module to a Retinal Scanner makes the player face disappear
+- Fix: The head of a sentry in Camouflage Mode can visually be retracted while the sentry is shooting at a target
+- Fix: Sentries always animate their head upwards when they first appear for a player
 
 --------------------------Changelog for v1.9.8 of SecurityCraft--------------------------